requirements.txt
# For metrics reporting
statsd>=3.3.0,<4.0.0
apache-airflow==2.6.3
pendulum==2.1.2
psycopg2-binary>=2.8.0
python-dateutil>=2.8.0
pydantic>=1.9.0,<2.0.0
